PongTournamentTest/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pypi_0
ale-py=0.10.1=pypi_0
alembic=1.14.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.2.0=py38hca03da5_0
appnope=0.1.2=py38hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h1a28f6b_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py38hca03da5_0
attrs=24.2.0=py38hca03da5_0
autorom=0.6.1=pypi_0
babel=2.11.0=py38hca03da5_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.3=py38hca03da5_0
bleach=4.1.0=pyhd3eb1b0_0
box2d-py=2.3.5=pypi_0
brotli-python=1.0.9=py38h313beb8_8
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.11.26=hca03da5_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py38hca03da5_0
cffi=1.17.1=py38h3eb5a62_0
chardet=5.2.0=pypi_0
charset-normalizer=3.4.0=pypi_0
chess=1.9.4=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.1.0=pypi_0
cmake=3.31.1=pypi_0
colorlog=6.9.0=pypi_0
comm=0.2.1=py38hca03da5_0
contextlib2=21.6.0=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
dataproperty=1.0.1=pypi_0
debugpy=1.6.7=py38h313beb8_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
dlib=19.24.6=cpu_py38h6f3644f_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=py38hca03da5_0
eval-type-backport=0.2.0=pypi_0
exceptiongroup=1.2.0=py38hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
farama-notifications=0.0.4=pypi_0
ffmpeg=1.4=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.55.3=pypi_0
fsspec=2024.10.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.36.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.68.1=pypi_0
gymnasium=1.0.0=pypi_0
h11=0.14.0=py38hca03da5_0
httpcore=1.0.2=py38hca03da5_0
httpx=0.27.0=py38hca03da5_0
huggingface-hub=0.26.5=pypi_0
huggingface-sb3=3.0=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
importlib_resources=6.4.0=py38hca03da5_0
ipykernel=6.29.5=py38hca03da5_0
ipython=8.12.2=py38hca03da5_0
jedi=0.19.1=py38hca03da5_0
jinja2=3.1.4=py38hca03da5_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.23.0=py38hca03da5_0
jsonschema-specifications=2023.7.1=py38hca03da5_0
jupyter-lsp=2.2.0=py38hca03da5_0
jupyter_client=7.4.9=py38hca03da5_0
jupyter_core=5.7.2=py38hca03da5_0
jupyter_events=0.10.0=py38hca03da5_0
jupyter_server=2.14.1=py38hca03da5_0
jupyter_server_terminals=0.4.4=py38hca03da5_1
jupyterlab=4.2.5=py38hca03da5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3=py38hca03da5_0
kiwisolver=1.4.7=pypi_0
libblas=3.9.0=25_osxarm64_openblas
libcblas=3.9.0=25_osxarm64_openblas
libcxx=19.1.5=ha82da77_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=25_osxarm64_openblas
libopenblas=0.3.28=openmp_hf332438_1
libpng=1.6.44=hc14010f_0
libsodium=1.0.18=h1a28f6b_0
libsqlite=3.47.2=h3f77e49_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.5=hdb05f8b_0
mako=1.3.8=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.6=py38hca03da5_0
mbstrdecoder=1.1.3=pypi_0
mdurl=0.1.2=pypi_0
mistune=2.0.4=py38hca03da5_0
ml-collections=0.1.1=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
multi-agent-ale-py=0.1.11=pypi_0
nbclient=0.8.0=py38hca03da5_0
nbconvert=7.16.4=py38hca03da5_0
nbformat=5.10.4=py38hca03da5_0
ncurses=6.5=h7bae524_1
nest-asyncio=1.6.0=py38hca03da5_0
networkx=3.1=pypi_0
notebook-shim=0.2.3=py38hca03da5_0
numpy=1.24.4=py38ha84db1f_0
oauthlib=3.2.2=pypi_0
open-spiel=1.4=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.4.0=h39f12f2_0
optuna=4.1.0=pypi_0
overrides=7.4.0=py38hca03da5_0
packaging=24.2=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathvalidate=3.2.1=pypi_0
pettingzoo=1.24.3=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.4.0=pypi_0
pip=24.2=py38hca03da5_0
pkgutil-resolve-name=1.3.10=py38hca03da5_1
platformdirs=3.10.0=py38hca03da5_0
proglog=0.1.12=pypi_0
prometheus_client=0.14.1=py38hca03da5_0
prompt-toolkit=3.0.43=py38hca03da5_0
protobuf=5.29.1=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pygame=2.3.0=pypi_0
pygments=2.18.0=pypi_0
pymunk=6.2.0=pypi_0
pyparsing=3.1.4=pypi_0
pysocks=1.7.1=py38hca03da5_0
pytablewriter=1.2.0=pypi_0
python=3.8.20=h7d35d02_2_cpython
python-dateutil=2.9.0post0=py38hca03da5_2
python-fastjsonschema=2.16.2=py38hca03da5_0
python-json-logger=2.0.7=py38hca03da5_0
python_abi=3.8=5_cp38
pytz=2024.2=pypi_0
pyyaml=6.0.2=py38h80987f9_0
pyzmq=25.1.2=py38h313beb8_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py38hca03da5_0
requests=2.32.3=py38hca03da5_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=py38hca03da5_0
rfc3986-validator=0.1.1=py38hca03da5_0
rich=13.9.4=pypi_0
rl-zoo3=2.4.0=pypi_0
rlcard=1.0.5=pypi_0
rpds-py=0.10.6=py38hf0e4da2_0
rsa=4.9=pypi_0
sb3-contrib=2.4.0=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.2=py38hca03da5_0
sentry-sdk=2.19.2=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.1.0=py38hca03da5_0
shimmy=2.0.0=pypi_0
six=1.17.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=py38hca03da5_0
soupsieve=2.5=py38hca03da5_0
sqlalchemy=2.0.36=pypi_0
stable-baselines3=2.4.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
supersuit=3.9.3=pypi_0
swig=4.3.0=pypi_0
sympy=1.13.3=pypi_0
tabledata=1.3.3=pypi_0
tcolorpy=0.1.6=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.17.1=py38hca03da5_0
tinycss2=1.2.1=py38hca03da5_0
tinyscaler=1.2.8=pypi_0
tk=8.6.13=h5083fa2_1
tomli=2.0.1=py38hca03da5_0
torch=2.4.1=pypi_0
tornado=6.4.1=py38h80987f9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=py38hca03da5_0
typepy=1.3.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=py38hca03da5_0
wandb=0.19.0=pypi_0
wasabi=1.1.3=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38hca03da5_1
websocket-client=1.8.0=py38hca03da5_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py38hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zipp=3.20.2=pypi_0
